Fix `Symbol::try_from_usize` truncating large integers
Currently, `Symbol::try_from_usize` only fails (returns `None`) when
the input `usize` is either exactly MAX value of the type or truncates to it.
This isn't an issue because symbols are generating sequentially, so
the panic will occur before there's a chance for this to happen, but it
is just weird code, there's no reason for it to be like that.